requirements.txt
# Base requirements

pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.0.0  # https://github.com/evansd/whitenoise
redis==4.2.2  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.50.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis


# PRECAUTION: avoid production dependencies that aren't in development
# Production requirements

gunicorn==20.1.0  # https://github.com/benoitc/gunicorn
psycopg2==2.9.3  # https://github.com/psycopg/psycopg2
hiredis==2.0.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django-anymail==8.5  # https://github.com/anymail/django-anymail

# ------------------------------------------------------------------------------
# New Packages for production
# ------------------------------------------------------------------------------
django-debug-toolbar
django_extensions
python-dotenv
django-filter
django-widget-tweaks
# ------------------------------------------------------------------------------
